// frontend/src/api/models/updateExtracurricularRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Inspirate Consulting API
 * API for the Inspirate Consulting application
 * OpenAPI spec version: 1.0.0
 */
import type { UpdateExtracurricularRequestStatus } from './updateExtracurricularRequestStatus';
import type { UpdateExtracurricularRequestType } from './updateExtracurricularRequestType';

export interface UpdateExtracurricularRequest {
  /** A URL to the JSON Schema for this object. */
  readonly $schema?: string;
  description?: string;
  end_date?: string;
  leadership_role?: string;
  /** @minLength 1 */
  name?: string;
  organization?: string;
  start_date?: string;
  status?: UpdateExtracurricularRequestStatus;
  type?: UpdateExtracurricularRequestType;
}